Cleaned up .multiSelectionResponsive code and documentation. (#684)
The second 'dim=' argument now only accepts a single dimension, either row or
column; this API change is allowed as the entire .multiSelectionResponsive
generic is only ever called in a single location across the iSEE codebase(s),
with a non-"none" output from .multiSelectionDimension(). It allows us to
greatly simplify the implementation of various methods.

Updated the documentation and signature to mention the new expectations. Also
added a default Panel method that conservatively requires a re-render upon any
multiple selection; subclasses can refine this for efficiency.